wrap loadlibrary call in try/catch block
The loadlibrary can fail due to various reasons. On Linux this includes
having a recent toolchain that supports newer C++ ABIs compared to the
libstdc++.so.* bundled in sys/os/glnxa64 with Matlab.

Since Matlab searches its own directories for bundled libs before the
system patchs it tries to use the older lib which lacks for support for
newer ABIs required by system libs that Matlab uses and doesn't bundle.
This causes errors similar to the following to be shown:

/opt/matlab/R2015a/bin/glnxa64/../../sys/os/glnxa64/libstdc++.so.6:
version `CXXABI_1.3.8' not found
(required by /usr/lib64/libicui18n.so.53)

The easiest fix is to override the libstdc++.so.6 symlink in
sys/os/glnxa64 to point to the system version instead of the bundled
Matlab version.

Catching the exception allows the user to take notice of it better and
provides better feedback to the app (in this case the filter wizard)
trying to load the library that things failed.

Signed-off-by: Tim Harder <timothy.harder@analog.com>